fix finance calculation

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 monthly_income = input("Enter your monthly income:")
 monthly_expenses = input("Enter your total monthly expenses:")
 monthly_savings = int(monthly_income) - int(monthly_expenses)
-projected_savings= (monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+projected_savings= (monthly_savings * 12 + float((monthly_savings * 12 * 0.05)))
 
 
 print ("Your monthly savings are $",monthly_savings)
